Refactor for wso2-extension to siddhi-io repo move
