+++ pom.xml
@@ -26,7 +26,7 @@
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
-    <groupId>org.wso2.extension.siddhi.execution.map</groupId>
+    <groupId>io.siddhi.extension.execution.map</groupId>
     <artifactId>siddhi-execution-map-parent</artifactId>
     <version>5.0.1-SNAPSHOT</version>
     <packaging>pom</packaging>
